feat: add testimonial card && chore: update testimonial content
